refactor and add unit tests
